1.6.0 ready

+++ src/main/java/org/overrun/glutils/timer/SystemTimer.java
@@ -50,7 +50,7 @@ public class SystemTimer extends AbstractTimer {
             passedNs = MAX_NS_PER_UPDATE;
         }
         fps = (float) (MAX_NS_PER_UPDATE / passedNs);
-        passedTime += (float) passedNs * timeScale * tps / (double) NS_PER_SECOND;
+        passedTime += (float) passedNs * timeScale * tps / (float) NS_PER_SECOND;
         ticks = (int) passedTime;
         if (ticks > MAX_TICKS_PER_UPDATE) {
             ticks = MAX_TICKS_PER_UPDATE;
